pkg/host/transport/ssh: make struct field name consistent with YAML
Not sure if Address is a best name for this field though. Host as a term
for descriping both hostname/fqdn and IP address seems like a better
choice, but I'd like to avoid the confusion with 'host` struct. Perhaps
we should rename it to samething else and change this one then.

We could use connection/host naming convention taken from Terraform:
https://www.terraform.io/docs/provisioners/connection.html#example-usage

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>