pkg/container/hostconfiguredcontainer: fix checking container status
If the container is removed manually and you call Status(),
container.Status will be set to nil, because the container is missing.

If you call Status() another time, it will return error, as you cannot
check status of non-existing container.

With this commit, Status() will not do anything, if container status is
missing, so it becomes idempotent.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>