Vagrantfile: add inline provisioning script
So Makefile does not have to take care of that and check on every run of
integration tests.

The implementation also slightly differs from the one in the Makefile,
as currently SSH authorized key changes we do there are not persisted
across the reboot, so this code adds public key to authorized_keys.d
directory and runs 'sudo update-ssh-keys' rather than adding directly to
authorized_keys file. This way the key is persisted across reboots.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>