pkg/container: batch stat operation and improve hostconfiguredcontainer
This commit modifies Runtime interface to allow batching stat operation,
which should reduce the number of operations we need to do on container
runtime and implements those changes for Docker runtime.

As making hostconfiguredcontainer compatible with this change would
produce a lot of duplicated code, this commit also adds few helper
functions to this struct, which acts as a wrappers for other actions,
which simplifies the overall code and makes it easier to test.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>