pkg/container/hostconfiguredcontainer: fix updating configuration status
a579063e64c7cd8247ec33e9ab6f036d302a361e prepared the code for reading
current content of the configuration files in batches, however, in the
updates status, paths with ConfigMountpoint included has been used,
which was making it basically useless for the user and it was triggering
contant updates of configuration files, even thought the real content
was correct.

This commit fixes that behavior by creating a map of original paths,
which is then used for restoring them in resulting map.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>