local-testing: initial commit
This commit adds local-testing directory, which links to e2e code. This
can be used to compile code and run Terraform locally, but still deploy
to remote machine.

That allows to quickly test independent pieces of the setup, where e2e
tests are more designed to make a general check if everything is still
fully functional.

local-testing also allows to test while developing charts, as it dumps
values.yaml files for each chart as a file.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>