pkg/container/runtime/docker: only pull container image if not present
Change default behavior to skip pulling image if it's already present on
the host. That drastically speeds up rolling updates to resources.

Given that the primary focus of that code is for deploying K8s, which is
critical infrastructure, pulling policy is currently not configurable,
as 'latest' tag should never really be used for it, to make sure
everything is compatible.

If there is a demand to implement 'pullPolicy: Always'-like behavior,
this can be done in the future.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>