terraform: fix inconsistent plan caused by changing config files
fc15390de5da11e423b72731f72e5bbc36469a34 attempted to resolve the issue
of Terraform reporting inconsistent plan, but it turns out it broke
updating the resource, as for Terraform, the changing value is fine'ish,
but new key appearing in the filed is not, hence we need to set the
fields to the empty value, if content of the config file is not known
(or empty). Then Terraform marks this fields as "Known After Apply".

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>