pkg/etcd: simplify handling member names
60f5c9b4f5c22cd8447601d14810aaa90d825eb9 attempted to fix handling
members with explicitly specified names, as this is what Terraform
resource is doing, but it introduced a bug, which broke removing
members, as member.getID function was never able to find the member ID
due to name mismatch and was always returning, that the member does not
exist anymore.

This commit simplifies how we handle the names. Member will now always
create containers with names with following format: 'etcd-<member
name>'.

Cluster, if member has no name specified explicitly, will simply use map
key as a name.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>