internal/utiltest: refactor to be useful also outside tests
In order to keep the validation logic for controlplane components
generic, we need to generate fake PKI for kubeconfig when we validate
kube-apiserver configuration, as it does not require kubeconfig to be
provided from the user.

This commit refactors GeneratePKI and adds GeneratePKIErr, which can be
used outside of unit tests. Because "return" linter does not does not
complain for multipe t.Fatal, but complains for multiple "return"
statements, we need to split GeneratePKI function into 2 smaller
functions. The benefit from that is increased readability, as now
functions are more specialized and easier to understand.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>