e2e/local-testing: prepare for enabling etcd RBAC
This commit adds new etcd client certificates to be generated in testing
environments, as with etcd RBAC enabled, each certificate's CN will
represent a separate user. The intention is to have following users:
- root - fully-privileged user for administrative actions
- kube-apiserver - user dedicated for kube-apiserver, also fully
  privileged for the time being because of
  https://github.com/etcd-io/etcd/issues/8458.
- prometheus - user for Prometheus to scrape etcd metrics

For local-testing, we also add rendering of few scripts, which can be
used for testing with etcdctl and to manually enable RBAC on etcd
cluster.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>