pkg/helm/release: fix version parameter being ignored for actions
This commit makes loadChart() use installClient() method, rather than
action.NewInstall() directly, to ensure, that the action client is
initialized correctly, mainly that it has the Version field set, what
installClient() do.

This makes sure, that when user defines which version of chart should be
installed, the setting is actually respected.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>