e2e: pin Kubernetes image and chart versions
This commit makes use of new 'version' argument for
flexkube_helm_release resources and enforces specific version of the
Flexkube charts to be used while running e2e tests, to make sure they
pass even if new version of the charts get released.

To fix some bugs, some charts use chart versions which either already
ship Kubernetes v1.19.0 or they ship Kubernetes below version v1.18.8,
so we pin the hyperkube image version to v1.18.8, which is a desired
version for the next release.

When running local tests, it is recommended to override the variables
default values with version '>0.0.0-0' to use latest possible version of
the charts.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>