pkg/container: improve handling missing host situation
Currently, if host running a container is lost, managing containers will
get stuck until lost container is removed manually from state.

This commit modifies the logic, so now in case where host which runs the
container cannot be reached, the error messages is added to the
container state and refreshing the state no longer returns error.

This allow user to inspect the state of a given container and decide
either to accept that that container is missing or abort and
investigate.

This should allow cleaning up the state gracefully even if all container
machines no longer exists.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>