Update golangci-lint to v1.33.0
This adds new linter which checks if all unit tests are run in parallel,
so this is added as well.

Some tests, e.g. ones which test using environment variables cannot run
in parallel, so the linter is explicitly disabled there with a comment.

Overall, it seems adding t.Parallel() cuts unit test execution time by
~2 seconds (~17s -> ~15s) and also makes sure that tests are independent
and exposes code which might be racey, so I think it make sense to add
it.

One rule from 'paralleltest' linter has been disabled, as it doesn't
make much sense. It has been reported here:
https://github.com/kunwardeep/paralleltest/issues/8

Also, now we add '--sort-output' flag to 'golangci-lint run' invokation,
so returned warnings are printed in a stable manner.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>