Remove unused -buildmode=exe flag
This flag comes from time when Arch Linux was using go-pie by default,
which IIRC has been printing some warnings, which this flag was hiding.

PIE is now default build mode on Linux and Windows, so it should be fine
not to use it, as there is no warnings printed either.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>