pkg/kubelet: remove /run/xtables.lock default mount
On OSes using nftables as iptables backend this file is no longer
necessary and won't exist, which may break creating kubelet container.

This commit removes this default mount, as right now there is no way to
do it on the user side. For users still not using nftables, mount must
be added manually.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>