e2e: fix failing integration tests with Docker
https://github.com/linuxserver/docker-baseimage-alpine/pull/85/commits/d2b11d7d0ca90a19da3bd7eb82eab078f7a24047
added tamper protection, which require custom configuration to be owned
by root, otherwise it is not executed.

Taking opportunity, we can actually get rid of Docker mods, as we apply
custom patches to configuration anyway, so then the openssh server
container does not need network for curl and jq anymore to pull
additional mods.

To workaround tampering check, we just use docker-compose to build
temporary images.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>